Fix script load order, add facebook service
