+++ static/js/greenwallet/init.js
@@ -1,11 +1,11 @@
 var app = require('./app');
 var raf = require('raf');
 
-app.run(['$templateCache', '$uibModal', function($templateCache, uid) {
+app.run(['$templateCache', '$uibModal', function ($templateCache, uid) {
   // we wait eeevvveerryyything loaded
-  process.nextTick(function() {
-    var content = $templateCache.get("uib/template/typeahead/typeahead-popup.html");
+  process.nextTick(function () {
+    var content = $templateCache.get('uib/template/typeahead/typeahead-popup.html');
     content = content.replace(/\)\}\}/g, ') }}');
-    $templateCache.put("uib/template/typeahead/typeahead-popup.html", content);
+    $templateCache.put('uib/template/typeahead/typeahead-popup.html', content);
   });
 }]);
